+++ src/kv/test/null_encryptor.h
@@ -7,7 +7,7 @@
 namespace kv
 {
   // NullTxEncryptor does not decrypt or verify integrity
-  class NullTxEncryptor : public kv::AbstractTxEncryptor
+  class NullTxEncryptor : public AbstractTxEncryptor
   {
   public:
     void encrypt(
@@ -15,7 +15,7 @@ namespace kv
       const std::vector<uint8_t>& additional_data,
       std::vector<uint8_t>& serialised_header,
       std::vector<uint8_t>& cipher,
-      kv::Version version,
+      const TxID& tx_id,
       bool is_snapshot = false) override
     {
       cipher = plain;
@@ -26,24 +26,20 @@ namespace kv
       const std::vector<uint8_t>& additional_data,
       const std::vector<uint8_t>& serialised_header,
       std::vector<uint8_t>& plain,
-      kv::Version version) override
+      Version version) override
     {
       plain = cipher;
       return true;
     }
-
-    void set_iv_id(size_t id) override {}
 
     size_t get_header_length() override
     {
       return 0;
     }
 
-    void update_encryption_key(
-      kv::Version version, const std::vector<uint8_t>& raw_ledger_key) override
-    {}
+    void disable_recovery() override {}
 
-    void rollback(kv::Version version) override {}
-    void compact(kv::Version version) override {}
+    void rollback(Version version) override {}
+    void compact(Version version) override {}
   };
 }