+++ src/node/genesis_gen.h
@@ -5,6 +5,7 @@
 #include "crypto/hash.h"
 #include "entities.h"
 #include "kv/tx.h"
+#include "ledger_secrets.h"
 #include "lua_interp/lua_interp.h"
 #include "lua_interp/lua_util.h"
 #include "members.h"
@@ -392,19 +393,26 @@ namespace ccf
       return true;
     }
 
-    void trust_node(NodeId node_id)
+    void trust_node(NodeId node_id, kv::Version latest_ledger_secret_seqno)
     {
       auto nodes_view = tx.get_view(tables.nodes);
       auto node_info = nodes_view->get(node_id);
-      if (node_info.has_value())
+
+      if (!node_info.has_value())
       {
-        node_info->status = NodeStatus::TRUSTED;
-        nodes_view->put(node_id, node_info.value());
+        throw std::logic_error(fmt::format("Node {} does not exist", node_id));
       }
-      else
+
+      if (node_info->status == NodeStatus::RETIRED)
       {
-        LOG_FAIL_FMT("Unknown node {} could not be trusted", node_id);
+        throw std::logic_error(fmt::format("Node {} is retired", node_id));
       }
+
+      node_info->status = NodeStatus::TRUSTED;
+      node_info->ledger_secret_seqno = latest_ledger_secret_seqno;
+      nodes_view->put(node_id, node_info.value());
+
+      LOG_INFO_FMT("Node {} is now {}", node_id, node_info->status);
     }
 
     auto get_last_signature()