+++ src/node/rpc/test/member_voting_test.cpp
@@ -222,7 +222,7 @@ auto get_cert(uint64_t member_id, tls::KeyPairPtr& kp_mem)
 }
 
 auto init_frontend(
-  NetworkTables& network,
+  NetworkState& network,
   GenesisGenerator& gen,
   StubNodeState& node,
   ShareManager& share_manager,
@@ -530,7 +530,8 @@ DOCTEST_TEST_CASE("Add new members until there are 7 then reject")
   constexpr auto n_new_members = 7;
   constexpr auto max_members = 8;
   NetworkState network;
-  network.ledger_secrets = std::make_shared<LedgerSecrets>();
+  NodeId node_id = 0;
+  network.ledger_secrets = std::make_shared<LedgerSecrets>(node_id);
   network.ledger_secrets->init();
   network.tables->set_encryptor(encryptor);
   auto gen_tx = network.tables->create_tx();
@@ -728,8 +729,7 @@ DOCTEST_TEST_CASE("Add new members until there are 7 then reject")
         create_signed_request(params, "ack", new_member->kp);
       const auto good_response =
         frontend_process(frontend, send_good_sig_req, new_member->cert);
-      DOCTEST_CHECK(good_response.status == HTTP_STATUS_OK);
-      DOCTEST_CHECK(parse_response_body<bool>(good_response));
+      DOCTEST_CHECK(good_response.status == HTTP_STATUS_NO_CONTENT);
 
       // (6) read own member status
       const auto read_status_req =
@@ -745,6 +745,9 @@ DOCTEST_TEST_CASE("Add new members until there are 7 then reject")
 DOCTEST_TEST_CASE("Accept node")
 {
   NetworkState network;
+  NodeId node_id = 0;
+  network.ledger_secrets = std::make_shared<LedgerSecrets>(node_id);
+  network.ledger_secrets->init();
   network.tables->set_encryptor(encryptor);
   auto gen_tx = network.tables->create_tx();
   GenesisGenerator gen(network, gen_tx);
@@ -772,7 +775,6 @@ DOCTEST_TEST_CASE("Accept node")
   gen.finalize();
   MemberRpcFrontend frontend(network, node, share_manager);
   frontend.open();
-  auto node_id = 0;
 
   // check node exists with status pending
   {
@@ -934,7 +936,7 @@ DOCTEST_TEST_CASE("Accept node")
 }
 
 ProposalInfo test_raw_writes(
-  NetworkTables& network,
+  NetworkState& network,
   GenesisGenerator& gen,
   StubNodeState& node,
   ShareManager& share_manager,
@@ -1473,6 +1475,9 @@ DOCTEST_TEST_CASE("Passing operator change" * doctest::test_suite("operator"))
   // Operator issues a proposal that is an operator change
   // and gets it through without member votes
   NetworkState network;
+  NodeId node_id = 0;
+  network.ledger_secrets = std::make_shared<LedgerSecrets>(node_id);
+  network.ledger_secrets->init();
   network.tables->set_encryptor(encryptor);
   auto gen_tx = network.tables->create_tx();
   GenesisGenerator gen(network, gen_tx);
@@ -1482,7 +1487,7 @@ DOCTEST_TEST_CASE("Passing operator change" * doctest::test_suite("operator"))
   auto new_ca = new_kp->self_sign("CN=new node");
   NodeInfo ni;
   ni.cert = new_ca;
-  auto node_id = gen.add_node(ni);
+  gen.add_node(ni);
 
   // Operating member, as indicated by member data
   const auto operator_cert = get_cert(0, kp);
@@ -1651,6 +1656,9 @@ DOCTEST_TEST_CASE(
   // Member proposes an operator change
   // A majority of members pass the vote
   NetworkState network;
+  NodeId node_id = 0;
+  network.ledger_secrets = std::make_shared<LedgerSecrets>(node_id);
+  network.ledger_secrets->init();
   network.tables->set_encryptor(encryptor);
   auto gen_tx = network.tables->create_tx();
   GenesisGenerator gen(network, gen_tx);
@@ -1692,7 +1700,6 @@ DOCTEST_TEST_CASE(
   const ccf::Script vote_for("return true");
   const ccf::Script vote_against("return false");
 
-  auto node_id = 0;
   {
     DOCTEST_INFO("Check node exists with status pending");
     const auto read_values =
@@ -1908,8 +1915,9 @@ DOCTEST_TEST_CASE("User data")
 
 DOCTEST_TEST_CASE("Submit recovery shares")
 {
-  NetworkState network(ConsensusType::CFT);
-  network.ledger_secrets = std::make_shared<LedgerSecrets>();
+  NetworkState network;
+  NodeId node_id = 0;
+  network.ledger_secrets = std::make_shared<LedgerSecrets>(node_id);
   network.ledger_secrets->init();
 
   ShareManager share_manager(network);
@@ -1951,10 +1959,10 @@ DOCTEST_TEST_CASE("Submit recovery shares")
 
     for (auto const& m : members)
     {
-      auto resp = parse_response_body<std::string>(
+      auto resp = parse_response_body<GetRecoveryShare::Out>(
         frontend_process(frontend, get_recovery_shares, m.second.first));
 
-      auto encrypted_share = tls::raw_from_b64(resp);
+      auto encrypted_share = tls::raw_from_b64(resp.encrypted_share);
       retrieved_shares[m.first] = m.second.second->unwrap(encrypted_share);
     }
   }
@@ -1962,8 +1970,9 @@ DOCTEST_TEST_CASE("Submit recovery shares")
   DOCTEST_INFO("Submit share before the service is in correct state");
   {
     MemberId member_id = 0;
-    const auto submit_recovery_share = create_text_request(
-      tls::b64_from_raw(retrieved_shares[member_id]), "recovery_share");
+    const auto submit_recovery_share = create_request(
+      SubmitRecoveryShare::In{tls::b64_from_raw(retrieved_shares[member_id])},
+      "recovery_share");
 
     check_error(
       frontend_process(
@@ -1994,8 +2003,9 @@ DOCTEST_TEST_CASE("Submit recovery shares")
     {
       auto bogus_recovery_share = retrieved_shares[m.first];
       bogus_recovery_share[0] = bogus_recovery_share[0] + 1;
-      const auto submit_recovery_share = create_text_request(
-        tls::b64_from_raw(bogus_recovery_share), "recovery_share");
+      const auto submit_recovery_share = create_request(
+        SubmitRecoveryShare::In{tls::b64_from_raw(bogus_recovery_share)},
+        "recovery_share");
 
       auto rep =
         frontend_process(frontend, submit_recovery_share, m.second.first);
@@ -2035,8 +2045,9 @@ DOCTEST_TEST_CASE("Submit recovery shares")
     size_t submitted_shares_count = 0;
     for (auto const& m : members)
     {
-      const auto submit_recovery_share = create_text_request(
-        tls::b64_from_raw(retrieved_shares[m.first]), "recovery_share");
+      const auto submit_recovery_share = create_request(
+        SubmitRecoveryShare::In{tls::b64_from_raw(retrieved_shares[m.first])},
+        "recovery_share");
 
       auto rep =
         frontend_process(frontend, submit_recovery_share, m.second.first);
@@ -2103,8 +2114,7 @@ DOCTEST_TEST_CASE("Number of active members with recovery shares limits")
       }
       else
       {
-        DOCTEST_CHECK(resp.status == HTTP_STATUS_OK);
-        DOCTEST_CHECK(parse_response_body<bool>(resp));
+        DOCTEST_CHECK(resp.status == HTTP_STATUS_NO_CONTENT);
       }
     }
   }
@@ -2118,8 +2128,7 @@ DOCTEST_TEST_CASE("Number of active members with recovery shares limits")
     gen.finalize();
     auto resp = activate(frontend, kp, cert);
 
-    DOCTEST_CHECK(resp.status == HTTP_STATUS_OK);
-    DOCTEST_CHECK(parse_response_body<bool>(resp));
+    DOCTEST_CHECK(resp.status == HTTP_STATUS_NO_CONTENT);
   }
 
   // Revert logging